Adjusted the code to use grids instead of block, in-line block, vertical-align, width, and margin
Created 6 video-previews and inserted them all into a <div> which represents the grid.

Reset the width to 100% in .thumbnail so the thumbnails fit into their columns to give more flexibility when you resize the page.

Remove the block property in .thumbnail as well as the width, inline-block, vertical-align and margin in the .video-preview to leave you with the 6 video-previews without gaps

Now add gaps using column-gap and row-gap